Prevent parallel blocks in 1-qubit circuits (#109)

Fixes #106
