+++ pytket/phir/api.py
@@ -72,7 +72,8 @@ def pytket_to_phir(
         # The function is called, but the output is just filled with 0s
         logger.debug("Performing placement and routing...")
     placed = place_and_route(shards, machine)
-    if machine:
+    # safety check: never run with parallelization on a 1 qubit circuit
+    if machine and len(circuit.qubits) > 1:
         phir_json = genphir_parallel(placed, machine)
     else:
         phir_json = genphir(placed, machine_ops=bool(machine))
