+++ tests/test_utils.py
@@ -45,6 +45,7 @@ class QasmFile(Enum):
     tk2_diff_angles = auto()
     rxrz = auto()
     classical_ordering = auto()
+    single_qubit_parallel_test = auto()
 
 
 class WatFile(Enum):
